docs: use fully qualified provider source (#19)
As the provider is not present on the OpenTofu registry, it's not usable by default as-is. The error message provided by OpenTofu is not very useful nor very googleable by people not familiar with terraform terminology. By fully qualifying the source (i.e. by including `registry.terraform.io`), the examples can be safely used by both OpenTofu and Terraform.